delete sim.py and rename sim_profile.py to sim.py
