+++ python/sim.py
@@ -1,70 +1,77 @@
 from dataclasses import dataclass
 from typing import List, Tuple
 
-import pid
+from pid import SimplePID
+from profile import piecewise_linear_setpoint
 
 
 @dataclass
 class ThermalModel:
-    target: float  # Target temperature (°C)
     ambient: float  # Ambient temperature (°C)
-    max_heating_rate: float  # Maximum heating rate (°C/s at 100% power)
+    max_heating_rate: float  # Max heating rate at 100% PWM (°C/s)
     cooling_coeff: float  # Cooling coefficient
 
 
-def simulate_tube_furnace(
-    config: ThermalModel, sim_time, dt: float = 0.1
+def clamp_pwm(output: float) -> float:
+    """Convert an unbounded PID output into a PWM fraction [0..1]."""
+    if output < 0:
+        return 0.0
+    elif output > 1.0:
+        return 1.0
+    else:
+        return output
+
+
+def simulate_profile(
+    profile: List[Tuple[float, float]], model: ThermalModel, sim_time: float, dt: float = 0.1
 ) -> Tuple[List[float], List[float], List[float], List[float], List[float]]:
     """
-    Simulate the temperature response of a tube furnace with PID control.
-
-    Args:
-        sim_time: Total simulation time (in seconds).
-        dt: Time step for the simulation (in seconds).
+    Simulate the furnace temperature using a piecewise linear setpoint profile
+    and a basic PID control that drives PWM from 0..1.
 
     Returns:
-        A tuple containing lists of simulation times, temperatures, PWM values,
-        PID outputs, and computed ramp rates.
+        times: time in seconds
+        temperatures: process temperature
+        setpoints: piecewise linear setpoint at each step
+        pwm_values: PWM fraction [0..1]
+        pid_outputs: raw PID outputs before clamping
     """
-    n_steps = int(sim_time / dt)
+    pid_controller = SimplePID(kp=0.5, ki=0.05, kd=1.0, imax=100.0)
+    pid_controller.reset()
 
-    times: List[float] = []
-    temperatures: List[float] = []
-    pwm_values: List[float] = []
-    pid_outputs: List[float] = []
-    ramp_rates: List[float] = []
+    # Start from ambient temperature
+    temperature = model.ambient
 
-    # Assume start temp = ambient temp
-    temperature = config.ambient
-
-    rs = pid.RampSoak(
-        kp=0.5,
-        ki=0.05,
-        kd=1.0,
-        imax=100.0,
-        ramp_up_limit=30,
-        ramp_down_limit=-30,
-        crossover_distance=10,
-        debug=False,
-    )
-    rs.set_target(config.target)
-    rs.reset_pid()
+    times = []
+    temperatures = []
+    setpoints = []
+    pwm_values = []
+    pid_outputs = []
 
     current_time = 0.0
-    for _ in range(n_steps):
-        output = rs.pid_step(temperature, dt)
-        pwm = pid.to_pwm(output)
-        # Simple model: heating is proportional to PWM, and cooling is proportional to the temperature excess over ambient.
-        heating = pwm * config.max_heating_rate
-        cooling = config.cooling_coeff * (temperature - config.ambient)
+    steps = int(sim_time / dt)
+
+    for _ in range(steps):
+        # 1) Retrieve the setpoint from the piecewise linear schedule
+        setpoint = piecewise_linear_setpoint(current_time, profile)
+        # 2) Let the PID compute its control output
+        output = pid_controller.compute(setpoint, temperature, dt)
+        # 3) Clamp output to [0..1] for PWM
+        pwm = clamp_pwm(output)
+
+        # 4) Compute net heating/cooling for one timestep
+        heating = pwm * model.max_heating_rate
+        cooling = model.cooling_coeff * (temperature - model.ambient)
+
         temperature += dt * (heating - cooling)
 
+        # Store data
         times.append(current_time)
         temperatures.append(temperature)
+        setpoints.append(setpoint)
         pwm_values.append(pwm)
         pid_outputs.append(output)
-        ramp_rates.append(rs.ramp_rate)
 
         current_time += dt
 
-    return times, temperatures, pwm_values, pid_outputs, ramp_rates
+    return times, temperatures, setpoints, pwm_values, pid_outputs
